another attempt at making 1D work nicely
The problem is that in 1D faces are 0D and so carry exactly no
useful information about which side of the element they may be
on.

To hack around that, this commit introduces a hacky operator
that hacks it. Less ambiguously, we just use the connection to
figure out which face we're on on set the normal to `+1` or `-1`
based on that